requirements.txt
requests
django-solo
